+++ examples/with-tailwindcss/tailwind.config.cjs
@@ -1,0 +1,7 @@
+module.exports = {
+	content: ['./src/**/*.{astro,html,js,jsx,svelte,ts,tsx,vue}'],
+	theme: {
+		extend: {},
+	},
+	plugins: [],
+};