+++ packages/astro/src/cli/index.ts
@@ -8,23 +8,24 @@ import yargs from 'yargs-parser';
 import { z } from 'zod';
 import { defaultLogDestination } from '../core/logger.js';
 import build from '../core/build/index.js';
+import add from '../core/add/index.js';
 import devServer from '../core/dev/index.js';
 import preview from '../core/preview/index.js';
 import { check } from './check.js';
-import { formatConfigError, loadConfig } from '../core/config.js';
-import { pad } from '../core/dev/util.js';
+import { loadConfig } from '../core/config.js';
+import { printHelp, formatErrorMessage, formatConfigErrorMessage } from '../core/messages.js';
+import { createSafeError } from '../core/util.js';
 
 type Arguments = yargs.Arguments;
-type CLICommand = 'help' | 'version' | 'dev' | 'build' | 'preview' | 'reload' | 'check';
+type CLICommand = 'help' | 'version' | 'add' | 'dev' | 'build' | 'preview' | 'reload' | 'check';
 
 /** Display --help flag */
-function printHelp() {
-	linebreak();
-	headline('astro', 'Futuristic web development tool.');
-	linebreak();
-	title('Commands');
-	table(
-		[
+function printAstroHelp() {
+	printHelp({
+		commandName: 'astro',
+		headline: 'Futuristic web development tool.',
+		commands: [
+			['add', 'Add an integration to your configuration.'],
 			['dev', 'Run Astro in development mode.'],
 			['build', 'Build a pre-compiled production-ready site.'],
 			['preview', 'Preview your build locally before deploying.'],
@@ -32,12 +33,7 @@ function printHelp() {
 			['--version', 'Show the version number and exit.'],
 			['--help', 'Show this help message.'],
 		],
-		{ padding: 28, prefix: '  astro ' }
-	);
-	linebreak();
-	title('Flags');
-	table(
-		[
+		flags: [
 			['--host [optional IP]', 'Expose server on network'],
 			['--config <path>', 'Specify the path to the Astro config file.'],
 			['--project-root <path>', 'Specify the path to the project root folder.'],
@@ -48,39 +44,7 @@ function printHelp() {
 			['--verbose', 'Enable verbose logging'],
 			['--silent', 'Disable logging'],
 		],
-		{ padding: 28, prefix: '  ' }
-	);
-
-	// Logging utils
-	function linebreak() {
-		console.log();
-	}
-
-	function headline(name: string, tagline: string) {
-		console.log(`  ${colors.bgGreen(colors.black(` ${name} `))} ${colors.green(`v${process.env.PACKAGE_VERSION ?? ''}`)} ${tagline}`);
-	}
-	function title(label: string) {
-		console.log(`  ${colors.bgWhite(colors.black(` ${label} `))}`);
-	}
-	function table(rows: [string, string][], opts: { padding: number; prefix: string }) {
-		const split = rows.some((row) => {
-			const message = `${opts.prefix}${' '.repeat(opts.padding)}${row[1]}`;
-			return message.length > process.stdout.columns;
-		});
-		for (const row of rows) {
-			row.forEach((col, i) => {
-				if (i === 0) {
-					process.stdout.write(`${opts.prefix}${colors.bold(pad(`${col}`, opts.padding - opts.prefix.length))}`);
-				} else {
-					if (split) {
-						process.stdout.write('\n    ');
-					}
-					process.stdout.write(colors.dim(col) + '\n');
-				}
-			});
-		}
-		return '';
-	}
+	});
 }
 
 /** Display --version flag */
@@ -93,15 +57,15 @@ async function printVersion() {
 
 /** Determine which command the user requested */
 function resolveCommand(flags: Arguments): CLICommand {
-	if (flags.version) {
-		return 'version';
-	} else if (flags.help) {
-		return 'help';
-	}
 	const cmd = flags._[2] as string;
+	if (cmd === 'add') return 'add';
+
+	if (flags.version) return 'version';
+	else if (flags.help) return 'help';
+
 	const supportedCommands = new Set(['dev', 'build', 'preview', 'check']);
 	if (supportedCommands.has(cmd)) {
-		return cmd as 'dev' | 'build' | 'preview' | 'check';
+		return cmd as CLICommand;
 	}
 	return 'help';
 }
@@ -110,11 +74,11 @@ function resolveCommand(flags: Arguments): CLICommand {
 export async function cli(args: string[]) {
 	const flags = yargs(args);
 	const cmd = resolveCommand(flags);
-	const projectRoot = flags.projectRoot || flags._[3];
+	const projectRoot = flags.projectRoot;
 
 	switch (cmd) {
 		case 'help':
-			printHelp();
+			printAstroHelp();
 			return process.exit(0);
 		case 'version':
 			await printVersion();
@@ -135,33 +99,37 @@ export async function cli(args: string[]) {
 
 	let config: AstroConfig;
 	try {
+		// Note: ideally, `loadConfig` would return the config AND its filePath
+		// For now, `add` has to resolve the config again internally
 		config = await loadConfig({ cwd: projectRoot, flags });
 	} catch (err) {
-		throwAndExit(err);
-		return;
+		return throwAndExit(err);
 	}
 
 	switch (cmd) {
+		case 'add': {
+			try {
+				const packages = flags._.slice(3) as string[];
+				return await add(packages, { cwd: projectRoot, flags, logging });
+			} catch (err) {
+				return throwAndExit(err);
+			}
+		}
 		case 'dev': {
 			try {
 				await devServer(config, { logging });
-
-				await new Promise(() => {}); // don’t close dev server
+				return await new Promise(() => {}); // lives forever
 			} catch (err) {
-				throwAndExit(err);
+				return throwAndExit(err);
 			}
-
-			return;
 		}
 
 		case 'build': {
 			try {
-				await build(config, { logging });
-				process.exit(0);
+				return await build(config, { logging });
 			} catch (err) {
-				throwAndExit(err);
+				return throwAndExit(err);
 			}
-			return;
 		}
 
 		case 'check': {
@@ -171,11 +139,10 @@ export async function cli(args: string[]) {
 
 		case 'preview': {
 			try {
-				await preview(config, { logging }); // this will keep running
+				return await preview(config, { logging }); // this will keep running
 			} catch (err) {
-				throwAndExit(err);
+				return throwAndExit(err);
 			}
-			return;
 		}
 
 		default: {
@@ -185,14 +152,11 @@ export async function cli(args: string[]) {
 }
 
 /** Display error and exit */
-function throwAndExit(err: any) {
+function throwAndExit(err: unknown) {
 	if (err instanceof z.ZodError) {
-		console.error(formatConfigError(err));
-	} else if (err.stack) {
-		const [mainMsg, ...stackMsg] = err.stack.split('\n');
-		console.error(colors.red(mainMsg) + '\n' + colors.dim(stackMsg.join('\n')));
+		console.error(formatConfigErrorMessage(err));
 	} else {
-		console.error(colors.red(err.toString() || err));
+		console.error(formatErrorMessage(createSafeError(err)));
 	}
 	process.exit(1);
 }